requirements.txt
tkinter
pillow
pynput
colorama
